Revert "Moving ArtifactTool utils to packaging common (#9383)"
This reverts commit ffe26c7146e1ff526c13b1120bed74360385bdef.
