+++ Tasks/Common/packaging-common/Tests/MockHelper.ts
@@ -2,12 +2,6 @@ import tmrm = require('vsts-task-lib/mock-run');
 	
 export function registerLocationHelpersMock(tmr: tmrm.TaskMockRunner) {
     const mockLocationUtils = {
-        getFeedUriFromBaseServiceUri: function (serviceUri: string, accesstoken: string) {
-            return serviceUri + "/feed"
-        },
-        getBlobstoreUriFromBaseServiceUri: function (serviceUri: string, accesstoken: string) {
-            return serviceUri + "/blobstore"
-        },
         getPackagingUris: function(input) {
             const collectionUrl: string = "https://vsts/packagesource";
             return {
