+++ Tasks/Common/packaging-common/package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@types/ltx": "^2.8.0",
     "@types/node": "^10.12.9",
-    "adm-zip": "^0.4.11",
     "azure-devops-node-api": "^6.6.0",
     "ini": "^1.3.4",
     "ip-address": "^5.8.9",
