+++ Tasks/UniversalPackagesV0/Common/Authentication.ts
