+++ Tasks/UniversalPackagesV0/task.json
@@ -10,7 +10,7 @@
     "version": {
         "Major": 0,
         "Minor": 147,
-        "Patch": 2
+        "Patch": 0
     },
     "runsOn": [
         "Agent",
